FEATURE : Added option to not initialize a pgdata directory.

+++ hitchpostgres/postgres_service.py
@@ -37,16 +37,14 @@ class PostgresUser(object):
 class PostgresService(Service):
     stop_signal = signal.SIGQUIT
 
-    def __init__(self, postgres_package, port=15432, encoding='UTF-8', locale='en_US', pgdata=None, users=None, databases=None, **kwargs):
-        #self.version = version
+    def __init__(self, postgres_package, port=15432, users=None, databases=None, initialize=True, **kwargs):
         self.postgres_package = postgres_package
         self.encoding = encoding
         self.locale = locale
         self.port = port
-        #self.pg_installation = PostgresInstallation() if None else postgres_installation
         self.users = users
         self.databases = databases
-        self.pgdata = pgdata
+        self.initialize = initialize
         checks.freeports([port, ])
         kwargs['log_line_ready_checker'] = lambda line: "database system is ready to accept connections" in line
         super(PostgresService, self).__init__(**kwargs)
@@ -85,29 +83,26 @@ class PostgresService(Service):
             return self._command
 
     def setup(self):
-        #self.log("Checking postgresql version...")
-        #version_output = self.subcommand(self.pg_installation.exec_postgres, "--version").run(check_output=True)
-        #if self.version not in version_output:
-            #raise HitchException("Postgres version needed is {}, output is: {}.".format(self.version, version_output))
         self.log("Initializing postgresql database...")
-        shutil.rmtree(self.pgdata, ignore_errors=True)
-        # TODO : Put back encoding + locale
-        self.subcommand(self.postgres_package.initdb, self.pgdata).run()
+        if self.initialize:
+            shutil.rmtree(self.pgdata, ignore_errors=True)
+            self.subcommand(self.postgres_package.initdb, self.pgdata).run()
 
     def poststart(self):
-        self.log("Creating users and databases...")
-        for user in self.users:
-            self.psql(
-                """create user {} with password '{}';""".format(user.username, user.password)
-            ).run()
-        for database in self.databases:
-            self.psql(
-                """create database {} with owner {};""".format(
-                    database.name, database.owner.username
-                )
-            ).run()
-            if database.dump is not None:
-                self.psql(database=database.name, filename=database.dump).run()
+        if self.initialize:
+            self.log("Creating users and databases...")
+            for user in self.users:
+                self.psql(
+                    """create user {} with password '{}';""".format(user.username, user.password)
+                ).run()
+            for database in self.databases:
+                self.psql(
+                    """create database {} with owner {};""".format(
+                        database.name, database.owner.username
+                    )
+                ).run()
+                if database.dump is not None:
+                    self.psql(database=database.name, filename=database.dump).run()
 
     def psql(self, command=None, database="template1", filename=None):
         """Run PSQL command."""
